requirements.txt
Flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0
revChatGPT==4.1.6
cryptography==40.0.1